Summarize a git repo

Producing a summary report of a repo

git_report(path = ".", output_file = NULL,
  output_format = "html_document", template = "gitsum::repo_summary_simple",
  input_file = NULL, directory = "gitsum", cached = FALSE,
  .libpath_index = 1)

Arguments

path
The path to the git directory one wants to create summaries for.

output_file
Output file. If NULL then a default based on the name of the input file is chosen.

output_format
R Markdown output format to convert to. Pass "all" to render all formats defined within the file. Pass the name of a format (e.g. "html_document") to render a single format or pass a vector of format names to render multiple formats. Alternatively you can pass an output format object; e.g. html_document(). If NULL is passed then the output format is the first one defined in the YAML metadata of the input file (defaulting to HTML if none is specified).

template
the template file to be used. Either one from a package (e.g. gitsum::repo_summary_simple), or a user defined template, which is then indicated with a path to this file including extension, i.e. /users/xzy/a_template.Rmd.

input_file
the name of the temporary file that contains log data on which the report is base on. If NULL, a log file is created according to path, otherwise, a file is read.

directory
A directory to which the reports should be stored.

cached
Whether or not the report should be cached with rmarkdown, so the graphs are writen to a file.

.libpath_index
index of the Library according to base::libPaths in which the template should be searched. The list with the available directories is returned invisibly for situations in which .libPaths() cannot be emulated.

Details

Other packages can define templates. They must be store an .Rmd template in inst/report_templates. The data passed to the report template is log, the parsed log table.