Obtain a detailed git log

This function returns a git log in a tabular format.

parse_log_detailed(path = ".", update_dump = TRUE)

parse_log_detailed_full_run(path = ".", na_to_zero = TRUE,
  file_name = NULL, commit_range = NULL)

Arguments

path
The path to the git directory one wants to create summaries for.

update_dump
Whether or not to update the dump in .gitsum after parsing.

na_to_zero
Whether some NA values should be converted to zero. See 'Details'.

file_name
The path to a raw log. If NULL, a raw log is created and read in and deleted after read-in, otherwise, an existing log is read.

commit_range
A string of the form "hash1..hash2" indicating the commit range to parse. NULL means all commits.

Value

A parsed git log as a nested tibble. Each line corresponds to a commit. The unnested column names are: short_hash, author_name, date, short_message, hash, left_parent, right_parent, author_email, weekday, month, monthday, time, year, timezone, message, description, total_files_changed, total_insertions, total_deletions, short_description, is_merge The nested columns contain more information on each commit. The column names are: changed_file, edits, insertions, deletions.

Details

Note that for merge commmits, the following columns are NA if the option na_to_zero is set to FALSE.: total_files_changed, total_insertions, total_deletions, changed_file, edits, deletions, insertions.

Note that for binary files, the following columns are 0: edits, deletions, insertions.

Functions

parse_log_detailed_full_run: In contrast to parse_log_detailed, this function does not read any history from the .gitum directory.

Warning

The number of edits, insertions, and deletions (on a file level) are based on git log --stat and the number of + and - in this log. The number of + and - may not sum up to the edits indicated as a scalar after "|" in git log --stat for commits with very many changed lines since for those, the + and - only indicate the relative share of insertions and edits. Therefore, parse_log_detailed_full_run() normalizes the insertions and deletions and rounds these after the normalization to achieve more consistent results. However, there is no guarantee that these numbers are always exact. The column is_exact indicates for each changed file within a commit whether the result is exact (which is the case if the normalizing constant was one).

See also

See parse_log_simple for a fast alternative with less information.