Use the latest name of a file in every log entry

Set the column changed_files of a log to the latest name of the file

set_changed_file_to_latest_name(log)

Arguments

log
An unnested detailed log.

Details

parse_log_detailed() parses a git log. However, changed file names are not corrected retroactively. Hence, if one wants to process file-specific commit data for one file across different names for that file, the column changed_files has to be unified. This is done with this function by updating the file names recursively and setting them to the latest name used. Filenames containing the sequence => are viewed as renamed files, so there will likely be problems when you apply this function to a git repository that contains (or contained) files with such files names. You may have to update the column changed_file manually.

Examples

library("magrittr")
gitsumlog %>%
  tidyr::unnest() %>%
  set_changed_file_to_latest_name() %>%
  add_line_history()

#> The following name changes were identified (6 in total):
#> ● man/{get_log.Rd => get_log_simple.Rd}
#> ● man/{parse_log.Rd => parse_log_one.Rd}
#> ● man/{get_pattern.Rd => get_pattern_multiple.Rd}
#> ● man/{get_log_regex.Rd => git_log_detailed.Rd}
#> ● man/{rmd_simple.Rd => git_report.Rd}
#> ● R/{gitsum.R => gitsum-package.R}

#> # A tibble: 193 x 27
#> # Groups:   changed_file [57]
#>    short_hash author_name  date                short_message  hash  left_parent
#>    <chr>      <chr>        <dttm>              <chr>          <chr> <chr>      
#>  1 243f       Lorenz Walt… 2017-05-19 11:23:23 initial commit 243f… <NA>       
#>  2 243f       Lorenz Walt… 2017-05-19 11:23:23 initial commit 243f… <NA>       
#>  3 243f       Lorenz Walt… 2017-05-19 11:23:23 initial commit 243f… <NA>       
#>  4 243f       Lorenz Walt… 2017-05-19 11:23:23 initial commit 243f… <NA>       
#>  5 243f       Lorenz Walt… 2017-05-19 11:23:23 initial commit 243f… <NA>       
#>  6 243f       Lorenz Walt… 2017-05-19 11:23:23 initial commit 243f… <NA>       
#>  7 243f       Lorenz Walt… 2017-05-19 11:23:23 initial commit 243f… <NA>       
#>  8 f8ee       Lorenz Walt… 2017-05-19 14:26:32 add log examp… f8ee… 243fcdb327…
#>  9 6328       Lorenz Walt… 2017-05-19 14:26:53 add parents    6328… f8ee8fb785…
#> 10 6328       Lorenz Walt… 2017-05-19 14:26:53 add parents    6328… f8ee8fb785…
#> # ... with 183 more rows, and 21 more variables: right_parent <chr>,
#> #   author_email <chr>, weekday <chr>, month <chr>, monthday <int>,
#> #   time <time>, year <int>, timezone <int>, message <chr>, description <chr>,
#> #   total_files_changed <int>, total_insertions <int>, total_deletions <int>,
#> #   short_description <chr>, is_merge <lgl>, changed_file <chr>, edits <int>,
#> #   insertions <dbl>, deletions <dbl>, lines_added <dbl>, current_lines <dbl>